Add support to multiple instances
The plugin works perfectly, but using IDs and counts (0, 1, 2, ...) makes this inflexible - for example, if you run more than one instance per page, it'll break - because two IDs fadeShow-slide0 will be created.

I modified the code a bit to use classes and point to elements inside the container, instead looking globally. This way, we can have as many instances as we want inside one page.